+++ src/main/java/com/br/springtesteautomatizado/models/PaymentProof.java
@@ -11,7 +11,7 @@ public class PaymentProof {
 
     @Id
     @GeneratedValue(strategy= GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     @Column(nullable = false)
     private LocalDate paymentDate;
     @ManyToOne
@@ -21,18 +21,18 @@ public class PaymentProof {
 
     public PaymentProof() {
     }
-    public PaymentProof(Integer id, LocalDate paymentDate, User user, BigDecimal amountPaid) {
+    public PaymentProof(Long id, LocalDate paymentDate, User user, BigDecimal amountPaid) {
         this.id = id;
         this.paymentDate = paymentDate;
         this.user = user;
         this.amountPaid = amountPaid;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
