+++ src/main/java/com/br/springtesteautomatizado/repositories/SaleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SaleRepository extends CrudRepository<Sale, Integer> {
+public interface SaleRepository extends CrudRepository<Sale, Long> {
 }
